backend/coffee_shop/products/serializers.py
from rest_framework.serializers import ModelSerializer

from products.models import Product


class ProductSerializer(ModelSerializer):
    class Meta:
        model = Product
        fields = ["name", "description", "price", "available", "creation_date_time"]
