backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-tailwind[reload]==3.8.0
crispy-tailwind==1.0.3
pillow==10.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==23.0.0
uvicorn==0.30.6
whitenoise[brotli]==6.7.0
